consolidated common methods to base Timer class, use duration constant instead of method, clarify README, tweak tests to make more sense
