+++ psalm.xml.dist
@@ -33,7 +33,7 @@
         </UnusedClass>
         <MissingThrowsDocblock>
             <errorLevel type="suppress">
-                <directory name="tests/Unit" />
+                <directory name="tests" />
             </errorLevel>
         </MissingThrowsDocblock>
     </issueHandlers>
